left, right, and bottom borders are 4x thiccer
The borders were changed from 1 px to 4 px. The extra 3 px use the
darker "background" color of the title bar and inactive border. The top
corners were given a larger chamfer, since I thought that looked better.
The bottom corners were given a chamfer, though it's not as big as the
chamfer for the top corners. It looked weird and pinched when it was as
wide as the top chamfers. The change in thiccness forced me to turn some
of the left-side files into separate XPMs rather than symlinks.