color modulation script; customizer list; readme
I wrote a short Bash script to recolor the theme by modulating all the
colors. This can be used to create a theme of arbitrary coloration. For
example, you could create a rusty orange theme by reducing the
brightness and saturation slightly and shifting the hue 15 percent.

I haven't tested the script yet. That will come in the next commit.